events: migrate severity to the OPC UA 1-1000 scale
Severity moves to the BaseEventType scale end to end (UInt16 1-1000,
0 invalid; OPC UA Part 5 §6.4.2): the EventSeverity constants rescale
({1,200,500,600,800,1000}), device events map to {1,600,800,1000}, the
wire projection emits UInt16 (UInt32 tolerated for the rollout window),
and the DataItemType_Severity config property joins the same scale —
its nodeset default becomes 10 and stored values rescale once via the
new config_migrations marker (s' = clamp(10*s, 1, 1000)). Historian
event tables rescale in the same gated step that adds the SourceName
column. Client thresholds follow the constants; profile-persisted
severityMin values rescale via the severityScale marker. The web app
already banded on 400/600/800 and now renders correctly. Phase 1 of
ADR 0005; redeploy tiers and client together.

+++ opcua_bridge/server_adapters_unittest.cpp
@@ -157,7 +157,7 @@ TEST(ServerAdapterTest, EventNotificationProjectsRealFieldValuesToOpcua) {
   ASSERT_EQ(event_fields->event_fields.size(), 3u);
   EXPECT_EQ(event_fields->event_fields[0].get<opcua::LocalizedText>(),
             opcua::LocalizedText{u"custom alarm"});
-  EXPECT_EQ(event_fields->event_fields[1].get<opcua::UInt32>(), 600u);
+  EXPECT_EQ(event_fields->event_fields[1].get<opcua::UInt16>(), 600u);
   // EventId is projected as ByteString per OPC UA Part 5 §6.4.2 BaseEventType,
   // https://reference.opcfoundation.org/Core/Part5/v105/docs/6.4.2
   EXPECT_EQ(event_fields->event_fields[2].get<opcua::ByteString>(),
